+++ mypyc/codegen/emitfunc.py
